# Avoid unnecessary construction of validator